Implement _input.c
Ported CPython input.c to uPy _input.c

Added select function for caller to use as _input.select, which
simulates select.select() in CPython

Added ioctl function for caller to use as _input.ioctl, which simulates
fnctl.ioctl() in CPython

Currently the buttons work now, with the change in ev3dev2 button.py,
changing from fnctl.ioctl to the ev_device.ioctl

So far this is only a quick hack. There are lots of things can be
improved. But hey, it works!